Add basic Signer interface, move types to location where they can be shared
