+++ apps/extension/src/extension/index.ts
@@ -1,4 +1,5 @@
 export * from "./ExtensionRequester";
 export * from "./ExtensionRouter";
 export * from "./guards";
+export * from "./Signer";
 export * from "./utils";
