+++ apps/extension/src/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./accounts";
